+++ src/PurchaseValidate.js
@@ -1,5 +1,4 @@
-const { DEFAULT, ERROR } = require("../utils/constant.js");
-const { Console } = require("@woowacourse/mission-utils");
+const { DEFAULT, ERROR } = require("./utils/constant.js");
 
 class PurchaseValudate {
   #input;
@@ -21,14 +20,9 @@ class PurchaseValudate {
   }
 
   check() {
-    if (this.checkNumber() || this.checkPositivNumber()) {
-      Console.close();
-      throw ERROR.PURCHASE_ERROR;
-    }
-    if (!this.checkUnit()) {
-      Console.close();
-      throw ERROR.PURCHASE_CHARGE;
-    }
+    if (this.checkNumber() || this.checkPositivNumber())
+      throw new Error(ERROR.PURCHASE_ERROR);
+    if (!this.checkUnit()) throw new Error(ERROR.PURCHASE_CHARGE);
   }
 
   getPurchaceAmount() {
